perf: move marked.use() calls to a singleton config module
